move filters in filtes.py
